feat: Restores connection on module, not startup
